CIV-8314 Moved the Flight Delay API test to steps_LRSmallSpec as we are testing for Small Claim with amount more than 1000
