+++ e2e/fixtures/events/evidenceUploadApplicant.js
@@ -172,15 +172,7 @@ module.exports = {
                   },
                   createdDatetime: '2023-02-06T13:11:52.466Z'
                 }
-              }],
-              witnessStatementFlag: 'do_not_show',
-              trialAuthorityFlag: 'do_not_show',
-              expertJointFlag: 'do_not_show',
-              witnessReferredStatementFlag: 'do_not_show',
-              expertReportFlag: 'do_not_show',
-              trialCostsFlag: 'do_not_show',
-              witnessSummaryFlag: 'do_not_show',
-              trialDocumentaryFlag: 'do_not_show'
+              }]
             }
           }
         };
